+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log (node-enumerable)
 
-## 3.8.0 (November 8th, 2017; methods)
+## 3.8.0 (November 7th, 2017; methods)
 
 * added `append()` and `appendArray()` methods
 * added `isEmpty()` method